pyproject.toml
[project]
name = "tg-bot-filter"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "aiogram>=3.22.0",
    "python-dotenv>=1.1.1",
]
